Refactor some of HttpConnection into HttpClient
